+++ src/rpn/src/Calculator.php
@@ -13,7 +13,6 @@ namespace Hyperf\Rpn;
 
 use Hyperf\Rpn\Exception\InvalidExpressionException;
 use Hyperf\Rpn\Exception\InvalidOperatorException;
-use Hyperf\Rpn\Exception\InvalidValueException;
 use Hyperf\Rpn\Exception\NotFoundException;
 use Hyperf\Rpn\Operator\AddOperator;
 use Hyperf\Rpn\Operator\DivideOperator;
@@ -50,7 +49,7 @@ class Calculator
     public function calculate(string $expression, array $bindings = [], int $scale = 0): string
     {
         $queue = new SplQueue();
-        $tags = $this->fromBindings(explode(' ', $expression), $bindings);
+        $tags = explode(' ', $expression);
         foreach ($tags as $tag) {
             if (! $this->isOperator($tag)) {
                 $queue->push($tag);
@@ -63,9 +62,6 @@ class Calculator
             $length = $operator->length();
             while (true) {
                 $value = $queue->pop();
-                if (! is_numeric($value)) {
-                    throw new InvalidValueException(sprintf('The value %s is invalid.', $value));
-                }
 
                 if ($length === null && $this->isOperator($value)) {
                     $queue->push($value);
@@ -79,7 +75,7 @@ class Calculator
                 }
             }
 
-            $queue->push($operator->execute(array_reverse($params), $scale));
+            $queue->push($operator->execute(array_reverse($params), $scale, $bindings));
         }
 
         if ($queue->count() !== 1) {