+++ src/rpn/src/Operator/AddOperator.php
@@ -18,8 +18,9 @@ class AddOperator extends Operator
         return '+';
     }
 
-    public function execute(array $parameters, int $scale): string
+    public function execute(array $parameters, int $scale, array $bindings = []): string
     {
+        $parameters = $this->fromBindings($parameters, $bindings);
         $parameters[] = $scale;
         return bcadd(...$parameters);
     }