+++ src/server/src/Listener/AfterWorkerStartListener.php
@@ -12,6 +12,7 @@ declare(strict_types=1);
 namespace Hyperf\Server\Listener;
 
 use Hyperf\Contract\StdoutLoggerInterface;
+use Hyperf\Engine\Constant\SocketType;
 use Hyperf\Event\Contract\ListenerInterface;
 use Hyperf\Framework\Event\AfterWorkerStart;
 use Hyperf\Server\Event\MainCoroutineServerStart;
@@ -56,13 +57,14 @@ class AfterWorkerStartListener implements ListenerInterface
                     switch ($type) {
                         case ServerInterface::SERVER_BASE:
                             $sockType = $server->type;
-                            // type of Swoole\Coroutine\Server is equal to SWOOLE_SOCK_UDP
-                            if ($server instanceof \Swoole\Coroutine\Server || in_array($sockType, [SWOOLE_SOCK_TCP, SWOOLE_SOCK_TCP6])) {
+                            // type of Swoole\Coroutine\Server is AF_INET which is equal to SWOOLE_SOCK_UDP
+                            if ($server instanceof \Swoole\Coroutine\Server || in_array($sockType, [SocketType::TCP, SocketType::TCP6])) {
                                 return 'TCP';
                             }
-                            if (in_array($sockType, [SWOOLE_SOCK_UDP, SWOOLE_SOCK_UDP6])) {
+                            if (in_array($sockType, [SocketType::UDP, SocketType::UDP6])) {
                                 return 'UDP';
                             }
+
                             return 'UNKNOWN';
                         case ServerInterface::SERVER_WEBSOCKET:
                             return 'WebSocket';