+++ src/swagger/publish/swagger.php
@@ -10,10 +10,13 @@ declare(strict_types=1);
  * @license  https://github.com/hyperf/hyperf/blob/master/LICENSE
  */
 return [
-    'enable' => env('APP_ENV', 'dev') !== 'prod',
+    'enable' => true,
     'port' => 9500,
-    'json' => BASE_PATH . '/openapi.json',
+    'json_dir' => BASE_PATH . '/storage/swagger',
     'html' => null,
     'url' => '/swagger',
     'auto_generate' => true,
+    'scan' => [
+        'paths' => null,
+    ],
 ];