+++ src/swagger/src/Listener/BootSwaggerListener.php
@@ -21,11 +21,6 @@ use Hyperf\HttpServer\Router\DispatcherFactory;
 use Hyperf\Server\Event;
 use Hyperf\Server\Server;
 use Hyperf\Swagger\Annotation as SA;
-use Hyperf\Swagger\Annotation\Get;
-use Hyperf\Swagger\Annotation\Head;
-use Hyperf\Swagger\Annotation\Patch;
-use Hyperf\Swagger\Annotation\Post;
-use Hyperf\Swagger\Annotation\Put;
 use Hyperf\Swagger\Generator;
 use Hyperf\Swagger\HttpServer;
 use Hyperf\Swagger\Util;
@@ -82,11 +77,13 @@ class BootSwaggerListener implements ListenerInterface
         // Init Router
         $factory = $this->container->get(DispatcherFactory::class);
         $annotations = [
-            Get::class,
-            Head::class,
-            Patch::class,
-            Post::class,
-            Put::class,
+            SA\Get::class,
+            SA\Head::class,
+            SA\Patch::class,
+            SA\Post::class,
+            SA\Put::class,
+            SA\Delete::class,
+            SA\Options::class,
         ];
 
         foreach ($annotations as $annotation) {
@@ -100,9 +97,9 @@ class BootSwaggerListener implements ListenerInterface
                 $classAnnotations = AnnotationCollector::getClassAnnotations($class);
                 $methodAnnotations = AnnotationCollector::getClassMethodAnnotation($class, $method);
 
-                $serverAnnotations = Util::findAnnotations($methodAnnotations, SA\Server::class);
+                $serverAnnotations = Util::findAnnotations($methodAnnotations, SA\HyperfServer::class);
                 if (! $serverAnnotations) {
-                    $serverAnnotations = Util::findAnnotations($classAnnotations, SA\Server::class);
+                    $serverAnnotations = Util::findAnnotations($classAnnotations, SA\HyperfServer::class);
                 }
 
                 $middlewareAnnotations = Util::findAnnotations($methodAnnotations, Middleware::class);
@@ -110,7 +107,7 @@ class BootSwaggerListener implements ListenerInterface
 
                 /** @var Operation $opera */
                 foreach ($annotation->toAnnotations() as $opera) {
-                    /** @var SA\Server $serverAnnotation */
+                    /** @var SA\HyperfServer $serverAnnotation */
                     foreach ($serverAnnotations as $serverAnnotation) {
                         $factory->getRouter($serverAnnotation->name)->addRoute(
                             [$opera->method],